requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
gunicorn>=21.2.0
SQLAlchemy>=2.0
asyncpg>=0.27.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.3.post2
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0

